Blog selection list scrolls and blogs are unchecked by default
I'm not completely happy with the implementation. I would prefer that the height of the scrolling pane was variable and would expand as needed until it ran out of vertical space, but right now you either have to set it to a flex pane which makes it take up all available vertical space, or set a fixed height.

fixes #28, #27

git-svn-id: https://webos.svn.wordpress.org/trunk@21 ee5cff3a-74c1-40a4-9fec-33ab0844ee87