changing to 3 column interface
It was using a SlidingPane inside of another SlidingPane which seemed to work fine with enyo 0.9 but started behaving very strangely with enyo 0.10/enyo 1.0

There are now fewer layers of components to go through to make things happen. wp.WordPress becomes the main controller between everything which works well since it is the "owner" of each of the wp.WordPressClient instances so it receives all of the event callbacks.

git-svn-id: https://webos.svn.wordpress.org/trunk@241 ee5cff3a-74c1-40a4-9fec-33ab0844ee87